Heavy cleanup - removed aggregated variable options, added CmvLatEquilibrate
New Process CmvLatEquilibrate
-class added to LateralExchangeABC.h header
-new file LatEquilibrate.cpp
-added LAT_EQUIL (CommonFunctions.cpp; RavenInclude.h)
-added :LateralEquilibrate command (ParseInput.cpp)

Deleted Aggregation
-removed SetAsAggregator, IsAggregatorGroup, _aAggregateSV members from HRUGroup (HRUGroups.cpp/.h)
-removed SetAggregatedVariable() routine from Model.cpp/.h
-removed aggregation check from ModelInitialize.cpp
-removed support for :AggregatedVariable command, added deprecation warning (ParseInput.cpp)
--automatically uses CmvLatEquilibrate process instead
-remvoed Aggregator code from Solvers.cpp

Removed MAX_STATE_VARS impact - is now maximum # of vars MANIPULABLE by one process
-Dynamic generation of sv array in Advection.cpp
-corrected _aStateVariIndices size to reference MAX_STATE_VAR_TYPES, not MAX_STATE_VARS (Model.cpp/.h)
-added MAX_STATE_VAR_TYPES constant

Minor cleanup (VStudio upgrade)
-(int) casting in ChannelXSection
-proper initialization of _nTotalLatConnections in Model.cpp
-removed unuseed nSoilLayers from CmvCappilaryRise

git-svn-id: https://chyms.nrc.gc.ca/svn/raven/trunk@329 9a3899ca-7adb-4d27-9b38-7f4ebb3123b0

+++ src/ChannelXSect.cpp
@@ -491,7 +491,7 @@ void CChannelXSect::GenerateRatingCurvesFromProfile()
   ExitGracefullyIf((maxe-mine)<REAL_SMALL,
                    "CChannelXSect::GenerateRatingCurvesFromProfile: profile survey points all have same elevation",BAD_DATA);
 
-  double dz=(maxe-mine)/(N-1);
+  double dz=(maxe-mine)/((int)(N)-1.0);
 
   int i=0;
   for (double z=mine;z<maxe+0.5*dz;z+=dz)
